Unity 2015.4.1f.1 Platformer Game
The game symbolizes the childhood and adolescence life of the main character. The main character is dreaming. It consists of 3 balls. The enemies in the game represent the people whose main character is having problems in real life. For example, in 1.level, the character is a child and the boss is an enemy mother. Because the main character has the most mother in real life. He is in high school at 2.level and is a boss enemy math teacher. In 3.level, the character is now the director of the company that is working in the business and the boss enemy. The biggest love is Lilly, the girls who attacked him at every level.